requirements.txt
Django==3.2.13
django-ckeditor==6.4.0
psycopg2-binary==2.9.3
Pillow==9.1.1
dj-database-url==0.5.0
gunicorn==20.1.0
whitenoise==6.1.0